fix(ci+security): resolve all Ruff, Sourcery, CodeQL, and Gemini review findings
Sourcery SQL injection (blocking):
- src/inference/conversation_store.py: replace f-string IN clause with
  safe string concatenation of ?-only placeholders (Sourcery compliant)

Gemini critical/high security fixes:
- api/users.py: _get_current_user delegates to api.auth._decode_token to
  honour the _revoked_jti set (fixes token revocation bypass)
- api.py: /mesh/stats and /mesh/quota now require authentication via
  Depends(get_current_user) — fixes information disclosure
- src/section7/web_scraper.py: fix CVE ID iteration bug (was iterating
  chars of string instead of appending the full CVE ID)
- src/section7/scheduler.py: run_full_scrape() via asyncio.to_thread to
  avoid blocking the event loop during network I/O
- src/observability/library_pipeline.py: release asyncio lock before
  network I/O; add _send_batch() helper called outside the lock
- src/utils/security_checks.py: guard os.getuid() with hasattr check
  (Windows compatibility); fix call to get_provider_health() (was
  check_provider_health()); add explanatory comments to bare except clauses
- workers/rate-limit-service-go/main.go: fix time.Duration truncation by
  multiplying float64 * float64(time.Second) before converting

CodeQL findings:
- workers/audit-service/worker.py: remove duplicate _router + require_internal_auth
- workers/health-aggregator/worker.py: remove duplicate _router + require_internal_auth
- workers/health-aggregator/worker.py: replace str(exc) with type(exc).__name__
  in health check response (no stack trace exposure)
- api/health.py: add noqa: F401 for model imports used as fallback types
- api/users.py: add explanatory comments to bare except clauses
- workers/the-void/rust_crypto/__init__.py: remove unused hmac import

All checks: ruff --select E,F,W,B clean.

+++ workers/rate-limit-service-go/main.go
@@ -56,7 +56,7 @@ func (b *bucket) check(cost int) (bool, int, time.Time) {
 	c := float64(cost)
 	if b.tokens >= c {
 		b.tokens -= c
-		resetAt := now.Add(time.Duration((b.maxTokens-b.tokens)/b.ratePerSec) * time.Second)
+		resetAt := now.Add(time.Duration((b.maxTokens-b.tokens)/b.ratePerSec * float64(time.Second)))
 		return true, int(b.tokens), resetAt
 	}
 	// Denied: tell caller when bucket will be full enough
